+++ storage/storage.go
@@ -4,7 +4,7 @@ import "wordfulness/types"
 
 type IStorage interface {
 	GetAllCourses() ([]*types.Course, error)
-	GetCourse(id int) (*types.Course, error)
+	GetCourse(id int64) (*types.Course, error)
 	CreateCourse(name string) error
-	DeleteCourse(id int) error
+	DeleteCourse(id int64) error
 }
